Introduce TIR visitor pattern, consolidate constant passes, rename rewrite.rs
- Add `optimize/visitor.rs` with `TirVisitor` trait that centralizes the
  ~200-line traversal boilerplate every pass previously duplicated.
  Passes now only implement their transformation logic.

- Consolidate constant optimization passes into `optimize/constants/` module:
  - propagation.rs (const_prop): GlobalVarGet → literal substitution
  - folding.rs (const_fold): compile-time arithmetic evaluation
  - global_promotion.rs: runtime globals → compile-time when folded
  - branch_prune.rs: extracted from dce.rs, dead branch elimination

- Rename rewrite.rs → select_lowering.rs for clarity (it only does
  if-else → wasm select lowering)

- All passes rewritten to use TirVisitor, eliminating manual traversal.
  Net reduction of ~800 lines with identical behavior (4650 E2E tests pass).
