+++ wado-compiler/src/optimize/dce.rs
@@ -2263,469 +2263,6 @@ fn remove_dead_global_sets_expr(expr: &mut TirExpr, used: &IndexSet<(String, Str
     }
 }
 
-// ──────────────────────────────────────────────────────────────────────────────
-// Constant branch pruning
-// ──────────────────────────────────────────────────────────────────────────────
-
-/// Eliminate dead branches where the `if` condition is a compile-time boolean literal.
-///
-/// **Statement-level (`TirStmtKind::If`):**
-/// - `if true  { A } [else { B }]` → inline A's statements
-/// - `if false { A }`              → remove entirely
-/// - `if false { A } else { B }`   → inline B's statements
-///
-/// **Expression-level (`TirExprKind::If`):**
-/// - `if true  { A } [else { B }]` → `Block(A)`
-/// - `if false { A }`              → `Unit`
-/// - `if false { A } else { B }`   → `Block(B)`
-pub fn prune_constant_branches(project: &mut Project) -> bool {
-    let mut changed = false;
-    for module in project.tir_modules.values_mut() {
-        for func_rc in &module.functions {
-            let mut func = func_rc.borrow_mut();
-            changed |= prune_branches_in_function(&mut func);
-        }
-    }
-    changed
-}
-
-fn prune_branches_in_function(func: &mut TirFunction) -> bool {
-    let Some(body) = &mut func.body else {
-        return false;
-    };
-    prune_branches_in_block(body)
-}
-
-fn prune_branches_in_block(block: &mut TirBlock) -> bool {
-    let mut changed = false;
-    for stmt in &mut block.stmts {
-        changed |= prune_branches_in_stmt(stmt);
-    }
-    // Eliminate dead statements: constant `if`, empty labeled blocks.
-    changed |= eliminate_dead_stmts(block);
-    changed
-}
-
-fn prune_branches_in_stmt(stmt: &mut TirStmt) -> bool {
-    match &mut stmt.kind {
-        TirStmtKind::Let { value, .. } => prune_branches_in_expr(value),
-        TirStmtKind::Expr(expr) => prune_branches_in_expr(expr),
-        TirStmtKind::Return { value } => value.as_mut().is_some_and(prune_branches_in_expr),
-        TirStmtKind::If {
-            condition,
-            then_block,
-            else_block,
-        } => {
-            let mut changed = prune_branches_in_expr(condition);
-            changed |= prune_branches_in_block(then_block);
-            if let Some(eb) = else_block {
-                changed |= prune_branches_in_block(eb);
-            }
-            changed
-        }
-        TirStmtKind::Loop { body } => prune_branches_in_block(body),
-        TirStmtKind::LabeledBlock { block, .. } => prune_branches_in_block(block),
-        TirStmtKind::IfLet {
-            scrutinee,
-            then_block,
-            else_block,
-            ..
-        } => {
-            let mut changed = prune_branches_in_expr(scrutinee);
-            changed |= prune_branches_in_block(then_block);
-            if let Some(eb) = else_block {
-                changed |= prune_branches_in_block(eb);
-            }
-            changed
-        }
-        TirStmtKind::Break { value, .. } => value.as_mut().is_some_and(prune_branches_in_expr),
-        TirStmtKind::Continue => false,
-        TirStmtKind::LetDestructure { value, .. } => prune_branches_in_expr(value),
-        TirStmtKind::TaskReturn { .. } => {
-            unreachable!("TaskReturn should be eliminated by synthesis before this phase")
-        }
-    }
-}
-
-fn prune_branches_in_expr(expr: &mut TirExpr) -> bool {
-    let mut changed = false;
-
-    // Recurse into sub-expressions first (bottom-up)
-    match &mut expr.kind {
-        TirExprKind::Binary { left, right, .. } => {
-            changed |= prune_branches_in_expr(left);
-            changed |= prune_branches_in_expr(right);
-        }
-        TirExprKind::Unary { expr: inner, .. }
-        | TirExprKind::FieldAccess { expr: inner, .. }
-        | TirExprKind::Cast { expr: inner, .. }
-        | TirExprKind::VariantTag { expr: inner }
-        | TirExprKind::VariantTest { expr: inner, .. }
-        | TirExprKind::VariantPayload { expr: inner, .. } => {
-            changed |= prune_branches_in_expr(inner);
-        }
-        TirExprKind::Assign { target, value } => {
-            changed |= prune_branches_in_expr(target);
-            changed |= prune_branches_in_expr(value);
-        }
-        TirExprKind::Index {
-            expr: inner, index, ..
-        } => {
-            changed |= prune_branches_in_expr(inner);
-            changed |= prune_branches_in_expr(index);
-        }
-        TirExprKind::Call { args, .. } => {
-            for arg in args {
-                changed |= prune_branches_in_expr(&mut arg.expr);
-            }
-        }
-        TirExprKind::CmRawCall { args, .. } => {
-            for arg in args {
-                changed |= prune_branches_in_expr(arg);
-            }
-        }
-        TirExprKind::MethodCall { receiver, args, .. } => {
-            changed |= prune_branches_in_expr(receiver);
-            for arg in args {
-                changed |= prune_branches_in_expr(&mut arg.expr);
-            }
-        }
-        TirExprKind::IndirectCall { callee, args } => {
-            changed |= prune_branches_in_expr(callee);
-            for arg in args {
-                changed |= prune_branches_in_expr(arg);
-            }
-        }
-        TirExprKind::ClosureToCanonical { functor, .. } => {
-            changed |= prune_branches_in_expr(functor);
-        }
-        TirExprKind::Block(block) | TirExprKind::LabeledBlock { block, .. } => {
-            changed |= prune_branches_in_block(block);
-        }
-        TirExprKind::If {
-            condition,
-            then_branch,
-            else_branch,
-        } => {
-            changed |= prune_branches_in_expr(condition);
-            changed |= prune_branches_in_block(then_branch);
-            if let Some(eb) = else_branch {
-                changed |= prune_branches_in_block(eb);
-            }
-        }
-        TirExprKind::Match { expr: inner, arms } => {
-            changed |= prune_branches_in_expr(inner);
-            for arm in arms {
-                if let Some(guard) = &mut arm.guard {
-                    changed |= prune_branches_in_expr(guard);
-                }
-                changed |= prune_branches_in_expr(&mut arm.body);
-            }
-        }
-        TirExprKind::StructLiteral { fields, .. } => {
-            for field in fields {
-                changed |= prune_branches_in_expr(&mut field.value);
-            }
-        }
-        TirExprKind::TupleLiteral { elements, .. } => {
-            for elem in elements {
-                changed |= prune_branches_in_expr(elem);
-            }
-        }
-        TirExprKind::VariantConstruct { payload, .. } => {
-            if let Some(p) = payload {
-                changed |= prune_branches_in_expr(p);
-            }
-        }
-        TirExprKind::Closure { body, .. } => {
-            changed |= prune_branches_in_expr(body);
-        }
-        TirExprKind::GlobalVarSet { value, .. } => {
-            changed |= prune_branches_in_expr(value);
-        }
-        TirExprKind::Switch {
-            scrutinee,
-            arms,
-            default,
-            ..
-        } => {
-            changed |= prune_branches_in_expr(scrutinee);
-            for arm in arms {
-                changed |= prune_branches_in_block(arm);
-            }
-            changed |= prune_branches_in_block(default);
-        }
-        // Leaf nodes
-        TirExprKind::Local { .. }
-        | TirExprKind::FuncRef { .. }
-        | TirExprKind::GlobalVarGet { .. }
-        | TirExprKind::Capture { .. }
-        | TirExprKind::IntLiteral { .. }
-        | TirExprKind::FloatLiteral { .. }
-        | TirExprKind::StringLiteral(_)
-        | TirExprKind::BytesLiteral(_)
-        | TirExprKind::BoolLiteral(_)
-        | TirExprKind::CharLiteral(_)
-        | TirExprKind::Null
-        | TirExprKind::Unit
-        | TirExprKind::EnumConstruct { .. } => {}
-        TirExprKind::TemplateString { .. } => {
-            unreachable!("TemplateString should be expanded before this phase")
-        }
-    }
-
-    // Prune expression-level `if` with constant boolean condition.
-    if let TirExprKind::If { condition, .. } = &expr.kind
-        && let TirExprKind::BoolLiteral(value) = condition.kind
-    {
-        let TirExprKind::If {
-            then_branch,
-            else_branch,
-            ..
-        } = std::mem::replace(&mut expr.kind, TirExprKind::Unit)
-        else {
-            unreachable!();
-        };
-        if value {
-            expr.kind = TirExprKind::Block(then_branch);
-        } else if let Some(else_blk) = else_branch {
-            expr.kind = TirExprKind::Block(else_blk);
-        }
-        // false without else: type is Unit, TirExprKind::Unit is already set
-        changed = true;
-    }
-
-    // Simplify `{ expr; }` → `expr` (single-expression unlabeled block)
-    if let TirExprKind::Block(block) = &expr.kind
-        && block.stmts.len() == 1
-        && let TirStmtKind::Expr(_) = &block.stmts[0].kind
-    {
-        let TirExprKind::Block(block) = std::mem::replace(&mut expr.kind, TirExprKind::Unit) else {
-            unreachable!();
-        };
-        let mut stmts = block.stmts;
-        let TirStmtKind::Expr(inner) = stmts.remove(0).kind else {
-            unreachable!();
-        };
-        *expr = inner;
-        changed = true;
-    }
-
-    // Simplify `label: { break label: val; }` → `val`
-    // This pattern arises from inlining `return expr;` → `break label: expr;`
-    if let TirExprKind::LabeledBlock { label, block, .. } = &expr.kind
-        && block.stmts.len() == 1
-        && let TirStmtKind::Break {
-            label: Some(brk_label),
-            ..
-        } = &block.stmts[0].kind
-        && brk_label == label
-    {
-        let TirExprKind::LabeledBlock { block, .. } =
-            std::mem::replace(&mut expr.kind, TirExprKind::Unit)
-        else {
-            unreachable!();
-        };
-        let mut stmts = block.stmts;
-        let TirStmtKind::Break { value, .. } = stmts.remove(0).kind else {
-            unreachable!();
-        };
-        if let Some(inner) = value {
-            *expr = inner;
-        }
-        // else: break without value → Unit is already set
-        changed = true;
-    }
-
-    // Simplify `[label:] { }` → `()` (empty block, with or without label)
-    if matches!(&expr.kind, TirExprKind::Block(b) | TirExprKind::LabeledBlock { block: b, .. } if b.stmts.is_empty())
-    {
-        expr.kind = TirExprKind::Unit;
-        changed = true;
-    }
-
-    changed
-}
-
-/// Eliminate dead statements from a block:
-/// - `if true { A } [else { B }]` → inline A's statements
-/// - `if false { A }` → remove
-/// - `if false { A } else { B }` → inline B's statements
-/// - `label: { }` (empty labeled block) → remove
-/// - `label: { stmts }` (unused label) → flatten stmts into parent
-fn eliminate_dead_stmts(block: &mut TirBlock) -> bool {
-    let dominated = |s: &TirStmt| {
-        matches!(
-            &s.kind,
-            TirStmtKind::If { condition, .. }
-                if matches!(condition.kind, TirExprKind::BoolLiteral(_))
-        ) || matches!(
-            &s.kind,
-            TirStmtKind::LabeledBlock { label, block }
-                if block.stmts.is_empty() || !block_has_break_to(label, block)
-        ) || matches!(
-            &s.kind,
-            TirStmtKind::Expr(e) if matches!(e.kind, TirExprKind::Unit | TirExprKind::Block(_))
-        ) || matches!(
-            &s.kind,
-            TirStmtKind::Expr(e) if matches!(&e.kind, TirExprKind::LabeledBlock { label, block, .. } if !block_has_break_to(label, block))
-        )
-    };
-    if !block.stmts.iter().any(dominated) {
-        return false;
-    }
-
-    let old_stmts = std::mem::take(&mut block.stmts);
-    for stmt in old_stmts {
-        // Constant `if` → inline taken branch or drop
-        if let TirStmtKind::If { ref condition, .. } = stmt.kind
-            && let TirExprKind::BoolLiteral(value) = condition.kind
-        {
-            let TirStmtKind::If {
-                then_block,
-                else_block,
-                ..
-            } = stmt.kind
-            else {
-                unreachable!();
-            };
-            if value {
-                block.stmts.extend(then_block.stmts);
-            } else if let Some(else_blk) = else_block {
-                block.stmts.extend(else_blk.stmts);
-            }
-            continue;
-        }
-        // Labeled block with unused label → flatten stmts into parent
-        if let TirStmtKind::LabeledBlock {
-            ref label,
-            block: ref inner,
-        } = stmt.kind
-            && !block_has_break_to(label, inner)
-        {
-            let TirStmtKind::LabeledBlock { block: inner, .. } = stmt.kind else {
-                unreachable!();
-            };
-            block.stmts.extend(inner.stmts);
-            continue;
-        }
-        // Unit expression → drop (side-effect free)
-        if let TirStmtKind::Expr(e) = &stmt.kind
-            && matches!(e.kind, TirExprKind::Unit)
-        {
-            continue;
-        }
-        // Void block expression → flatten stmts into parent.
-        // Handles both `Expr(Block { ... })` and `Expr(LabeledBlock { unused label, ... })`.
-        // The latter arises from inlining void functions whose label has no break.
-        if let TirStmtKind::Expr(e) = &stmt.kind
-            && matches!(e.kind, TirExprKind::Block(_))
-        {
-            let TirStmtKind::Expr(e) = stmt.kind else {
-                unreachable!();
-            };
-            let TirExprKind::Block(inner) = e.kind else {
-                unreachable!();
-            };
-            block.stmts.extend(inner.stmts);
-            continue;
-        }
-        if let TirStmtKind::Expr(e) = &stmt.kind
-            && matches!(&e.kind, TirExprKind::LabeledBlock { label, block, .. } if !block_has_break_to(label, block))
-        {
-            let TirStmtKind::Expr(e) = stmt.kind else {
-                unreachable!();
-            };
-            let TirExprKind::LabeledBlock { block: inner, .. } = e.kind else {
-                unreachable!();
-            };
-            block.stmts.extend(inner.stmts);
-            continue;
-        }
-        block.stmts.push(stmt);
-    }
-    true
-}
-
-/// Check if any `break` statement in the block targets the given label.
-fn block_has_break_to(label: &str, block: &TirBlock) -> bool {
-    block.stmts.iter().any(|s| stmt_has_break_to(label, s))
-}
-
-fn stmt_has_break_to(label: &str, stmt: &TirStmt) -> bool {
-    match &stmt.kind {
-        TirStmtKind::Break { label: Some(l), .. } => l == label,
-        TirStmtKind::Let { value, .. } | TirStmtKind::LetDestructure { value, .. } => {
-            expr_has_break_to(label, value)
-        }
-        TirStmtKind::Expr(expr) => expr_has_break_to(label, expr),
-        TirStmtKind::If {
-            condition,
-            then_block,
-            else_block,
-        } => {
-            expr_has_break_to(label, condition)
-                || block_has_break_to(label, then_block)
-                || else_block
-                    .as_ref()
-                    .is_some_and(|b| block_has_break_to(label, b))
-        }
-        TirStmtKind::Loop { body } | TirStmtKind::LabeledBlock { block: body, .. } => {
-            block_has_break_to(label, body)
-        }
-        TirStmtKind::IfLet {
-            scrutinee,
-            then_block,
-            else_block,
-            ..
-        } => {
-            expr_has_break_to(label, scrutinee)
-                || block_has_break_to(label, then_block)
-                || else_block
-                    .as_ref()
-                    .is_some_and(|b| block_has_break_to(label, b))
-        }
-        TirStmtKind::Return { value } => {
-            value.as_ref().is_some_and(|v| expr_has_break_to(label, v))
-        }
-        _ => false,
-    }
-}
-
-fn expr_has_break_to(label: &str, expr: &TirExpr) -> bool {
-    match &expr.kind {
-        TirExprKind::Block(block) | TirExprKind::LabeledBlock { block, .. } => {
-            block_has_break_to(label, block)
-        }
-        TirExprKind::If {
-            condition,
-            then_branch,
-            else_branch,
-        } => {
-            expr_has_break_to(label, condition)
-                || block_has_break_to(label, then_branch)
-                || else_branch
-                    .as_ref()
-                    .is_some_and(|b| block_has_break_to(label, b))
-        }
-        TirExprKind::Match { expr, arms } => {
-            expr_has_break_to(label, expr)
-                || arms.iter().any(|arm| expr_has_break_to(label, &arm.body))
-        }
-        TirExprKind::Switch {
-            scrutinee,
-            arms,
-            default,
-            ..
-        } => {
-            expr_has_break_to(label, scrutinee)
-                || arms.iter().any(|arm| block_has_break_to(label, arm))
-                || block_has_break_to(label, default)
-        }
-        _ => false,
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
