+++ wado-compiler/src/optimize/select_lowering.rs
@@ -1,0 +1,125 @@
+//! Select lowering optimization for Wado TIR
+//!
+//! Post-optimization rewrite that converts simple `if cond { a } else { b }` expressions
+//! to `builtin::select(cond, a, b)`, which emits the branchless Wasm `select` instruction.
+//! Both branches must be pure (no side effects, no traps) since `select` evaluates both
+//! operands eagerly.
+
+use crate::name::ModuleSource;
+use crate::project::Project;
+use crate::tir::{
+    CallArg, FunctionRef, MonomorphInfo, TirBlock, TirExpr, TirExprKind, TirStmtKind, TypeId,
+    TypeTable,
+};
+
+use super::visitor::{walk_expr, TirVisitor};
+
+/// Run select lowering on all functions.
+pub fn select_lowering(project: &mut Project) {
+    let mut visitor = SelectLoweringVisitor;
+    for module in project.tir_modules.values_mut() {
+        for func_rc in &module.functions {
+            let mut func = func_rc.borrow_mut();
+            if let Some(ref mut body) = func.body {
+                visitor.visit_block(body);
+            }
+        }
+    }
+}
+
+struct SelectLoweringVisitor;
+
+impl TirVisitor for SelectLoweringVisitor {
+    fn visit_expr(&mut self, expr: &mut TirExpr) -> bool {
+        let mut changed = false;
+        match &mut expr.kind {
+            TirExprKind::If {
+                condition,
+                then_branch,
+                else_branch,
+            } => {
+                if let Some(select_call) = try_lower_to_select(
+                    condition,
+                    then_branch,
+                    else_branch,
+                    expr.type_id,
+                    expr.span,
+                ) {
+                    *expr = select_call;
+                    changed = true;
+                    changed |= self.visit_expr(expr);
+                    return changed;
+                }
+            }
+            _ => {}
+        }
+        changed |= walk_expr(self, expr);
+        changed
+    }
+}
+
+fn is_select_eligible_expr(expr: &TirExpr) -> bool {
+    matches!(
+        &expr.kind,
+        TirExprKind::Local { .. }
+            | TirExprKind::IntLiteral { .. }
+            | TirExprKind::FloatLiteral { .. }
+            | TirExprKind::BoolLiteral(_)
+            | TirExprKind::CharLiteral(_)
+    )
+}
+
+fn try_select_value(block: &TirBlock) -> Option<&TirExpr> {
+    if block.stmts.len() != 1 {
+        return None;
+    }
+    if let TirStmtKind::Expr(expr) = &block.stmts[0].kind
+        && is_select_eligible_expr(expr)
+    {
+        return Some(expr);
+    }
+    None
+}
+
+fn try_lower_to_select(
+    condition: &TirExpr,
+    then_branch: &TirBlock,
+    else_branch: &Option<TirBlock>,
+    result_type: TypeId,
+    span: crate::token::Span,
+) -> Option<TirExpr> {
+    let else_block = else_branch.as_ref()?;
+
+    if result_type == TypeTable::UNIT {
+        return None;
+    }
+
+    let true_val = try_select_value(then_branch)?;
+    let false_val = try_select_value(else_block)?;
+
+    let func_ref = FunctionRef {
+        module_source: ModuleSource::builtin(),
+        name: "select".to_string(),
+        monomorph_info: Some(MonomorphInfo {
+            generic_name: "select".to_string(),
+            type_args: vec![result_type],
+            is_blanket: false,
+        }),
+        method_info: None,
+        is_cm_adapter: false,
+    };
+
+    Some(TirExpr::new(
+        TirExprKind::Call {
+            func: func_ref,
+            type_args: vec![result_type],
+            args: vec![
+                CallArg::new(condition.clone(), false),
+                CallArg::new(true_val.clone(), false),
+                CallArg::new(false_val.clone(), false),
+            ],
+        },
+        result_type,
+        span,
+    ))
+}
